refactor(datetime): merge datetime module into country_metadata
Move all datetime functions (FormatDateTime, GetRelativeTime, GetTimeZoneByCountry, GetWeekdays) and their types/helpers into the country_metadata package. Delete the datetime module files.
